external_agents/treasury/requirements.txt
# Treasury Agent Dependencies
# Install: pip install -r requirements.txt

# Core dependencies
fredapi==0.5.2              # We'll use FRED API for Treasury data (most reliable)
requests==2.31.0            # HTTP requests (backup for Treasury.gov)
aiohttp==3.9.1              # Async HTTP requests for parallel API calls
pandas==2.2.0               # Data manipulation
numpy==1.26.3               # Numerical operations
python-dotenv==1.0.0        # Environment variable management

# ADK and A2A
google-adk>=0.1.0           # Agent Development Kit

# Utilities
cachetools==5.3.2           # Response caching
python-dateutil==2.8.2      # Date handling
scipy==1.11.4               # Statistical functions (for yield curve fitting)

# Optional: Development/Testing
pytest==7.4.4               # Testing framework
pytest-asyncio==0.23.3      # Async testing support

# Note: Treasury yield data is available via FRED API
# TIPS data also available via FRED
# We'll use FRED as the primary source for reliability
